refactor: reorganizando toda a estrutura e separando camadas (API, Database, Domain, Ioc, Shared)
