+++ CorePay.API.Application/Interfaces/ITransactionService.cs
@@ -1,0 +1,10 @@
+﻿using CorePayAPI.DTOs.Requests;
+using CorePayAPI.DTOs.Responses;
+
+namespace CorePay.API.Application.Interfaces
+{
+    public interface ITransactionService
+    {
+        Task<Response> TransferMoneyAsync(TransferRequest request);
+    }
+}
