+++ CorePay.API.Application/Interfaces/IUserService.cs
@@ -1,0 +1,9 @@
+﻿using CorePay.API.Domain.Entities;
+
+namespace CorePay.API.Application.Interfaces
+{
+    public interface IUserService
+    {
+        User ConsultUser(int userId);
+    }
+}
