+++ CorePay.API.Database/EntitiesConfigurations/UserEntityConfigurations.cs
@@ -1,0 +1,30 @@
+﻿using CorePay.API.Domain.Entities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace CorePayAPI.EntitiesConfigurations
+{
+    public class UserEntityConfigurations : IEntityTypeConfiguration<User>
+    {
+        public void Configure(EntityTypeBuilder<User> builder)
+        { 
+        builder.ToTable("User");
+                    builder.HasKey(e => e.Id);
+                    builder.Property(e => e.Id).ValueGeneratedOnAdd();
+
+            builder.Property(e => e.Name)
+                                .IsRequired()
+                                .HasMaxLength(100);
+
+            builder.Property(e => e.Balance)
+                                 .IsRequired();
+
+
+            builder.HasData(
+                    new User(1, "Matheus", 1000),
+                    new User(2, "Amanda", 10)
+);
+
+       }
+    }
+}
