+++ CorePay.API.Domain/Entities/Entity.cs
@@ -1,0 +1,9 @@
+﻿using CorePay.API.Domain.Interfaces;
+
+namespace CorePayAPI.Entities
+{
+    public class Entity : IEntity
+    {
+        public int Id { get; set; }
+    }
+}
