+++ CorePay.API.Domain/Interfaces/IEntity.cs
@@ -1,0 +1,7 @@
+﻿namespace CorePay.API.Domain.Interfaces
+{
+    public interface IEntity
+    {
+        public int Id { get; set; }
+    }
+}
