+++ CorePay.API.Domain/Interfaces/ITransferRepository.cs
@@ -1,0 +1,10 @@
+﻿using CorePay.API.Domain.Entities;
+
+namespace CorePayAPI.Repository.Interface
+{
+    public interface ITransferRepository
+    {
+        Task<User> GetUserByIdAsync(int id);
+        Task SaveChangesAsync();
+    }
+}
