+++ CorePay.API.Domain/Interfaces/IUserRepository.cs
@@ -1,0 +1,10 @@
+﻿using CorePay.API.Domain.Entities;
+
+namespace CorePayAPI.Repository.Interface
+{
+    public interface IUserRepository
+    {
+        User ConsultUser(int UserId);
+
+    }
+}
