+++ CorePay.API.Shared/Configuration/DbConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace CorePay.API.Shared.Configuration
+{
+    public class DbConfig
+    {
+        public string ConnectionString { get; set; }
+    }
+}
