+++ CorePayAPI.sln
@@ -3,7 +3,36 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35208.52
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CorePayAPI", "CorePayAPI\CorePayAPI.csproj", "{E430B41F-797E-4808-9A96-17E990B7961C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorePayAPI", "CorePayAPI\CorePayAPI.csproj", "{E430B41F-797E-4808-9A96-17E990B7961C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "API", "API", "{59EDBBA3-A497-434F-AE41-F9D54393B9B4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{8A5B62E6-39F7-4DD9-8793-165F6DBAE8BD}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Database", "Database", "{DBBEA4A7-7F6D-49FD-B104-FE6D86213169}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Ioc", "Ioc", "{6280985B-5414-4D44-AF62-5E9566366884}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{DF6971F4-FF20-4796-9B21-2D6AC428436B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Application", "Application", "{ED189B0C-D7DF-4E4C-AAE5-325FE63E0955}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain", "Domain", "{82E30B05-46A5-4685-BD70-F269F02E3D33}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorePay.API.Database", "CorePay.API.Database\CorePay.API.Database.csproj", "{B3D86631-CEAC-4DD7-956C-E91AFEDCD1FF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorePay.API.Ioc", "CorePay.IOC\CorePay.API.Ioc.csproj", "{D5899C72-F410-482B-B790-B3DAC8111F13}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SolutionsItems", "SolutionsItems", "{A12F6CD6-5BC2-4B70-BE43-C13009F772D9}"
+	ProjectSection(SolutionItems) = preProject
+		Migration.txt = Migration.txt
+	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorePay.API.Domain", "CorePay.API.Domain\CorePay.API.Domain.csproj", "{3DA1B2AF-2B66-4776-B7BA-724989AC8B6C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorePay.API.Application", "CorePay.API.Application\CorePay.API.Application.csproj", "{B2EE7A44-07BD-40A5-8244-B20CD5EA5661}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CorePay.API.Shared", "CorePay.API.Shared\CorePay.API.Shared.csproj", "{6D5BD0FB-56B8-41D7-B7CC-1D667FC348D4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,9 +44,39 @@ Global
 		{E430B41F-797E-4808-9A96-17E990B7961C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E430B41F-797E-4808-9A96-17E990B7961C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E430B41F-797E-4808-9A96-17E990B7961C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B3D86631-CEAC-4DD7-956C-E91AFEDCD1FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B3D86631-CEAC-4DD7-956C-E91AFEDCD1FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B3D86631-CEAC-4DD7-956C-E91AFEDCD1FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B3D86631-CEAC-4DD7-956C-E91AFEDCD1FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5899C72-F410-482B-B790-B3DAC8111F13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5899C72-F410-482B-B790-B3DAC8111F13}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5899C72-F410-482B-B790-B3DAC8111F13}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5899C72-F410-482B-B790-B3DAC8111F13}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3DA1B2AF-2B66-4776-B7BA-724989AC8B6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3DA1B2AF-2B66-4776-B7BA-724989AC8B6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3DA1B2AF-2B66-4776-B7BA-724989AC8B6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3DA1B2AF-2B66-4776-B7BA-724989AC8B6C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2EE7A44-07BD-40A5-8244-B20CD5EA5661}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2EE7A44-07BD-40A5-8244-B20CD5EA5661}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2EE7A44-07BD-40A5-8244-B20CD5EA5661}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2EE7A44-07BD-40A5-8244-B20CD5EA5661}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6D5BD0FB-56B8-41D7-B7CC-1D667FC348D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6D5BD0FB-56B8-41D7-B7CC-1D667FC348D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6D5BD0FB-56B8-41D7-B7CC-1D667FC348D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6D5BD0FB-56B8-41D7-B7CC-1D667FC348D4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{E430B41F-797E-4808-9A96-17E990B7961C} = {59EDBBA3-A497-434F-AE41-F9D54393B9B4}
+		{ED189B0C-D7DF-4E4C-AAE5-325FE63E0955} = {8A5B62E6-39F7-4DD9-8793-165F6DBAE8BD}
+		{82E30B05-46A5-4685-BD70-F269F02E3D33} = {8A5B62E6-39F7-4DD9-8793-165F6DBAE8BD}
+		{B3D86631-CEAC-4DD7-956C-E91AFEDCD1FF} = {DBBEA4A7-7F6D-49FD-B104-FE6D86213169}
+		{D5899C72-F410-482B-B790-B3DAC8111F13} = {6280985B-5414-4D44-AF62-5E9566366884}
+		{3DA1B2AF-2B66-4776-B7BA-724989AC8B6C} = {82E30B05-46A5-4685-BD70-F269F02E3D33}
+		{B2EE7A44-07BD-40A5-8244-B20CD5EA5661} = {ED189B0C-D7DF-4E4C-AAE5-325FE63E0955}
+		{6D5BD0FB-56B8-41D7-B7CC-1D667FC348D4} = {DF6971F4-FF20-4796-9B21-2D6AC428436B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4A634634-7382-4B4C-BDB2-8E605E6D106D}
