+++ CorePayAPI/Controllers/TransactionController.cs
@@ -1,5 +1,5 @@
-﻿using CorePayAPI.DTOs.Requests;
-using CorePayAPI.Services.Interface;
+﻿using CorePay.API.Application.Interfaces;
+using CorePayAPI.DTOs.Requests;
 using Microsoft.AspNetCore.Mvc;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
