--- CorePayAPI/Entities/Entity.cs
@@ -1,9 +1,0 @@
-﻿using CorePayAPI.Entities.Interface;
-
-namespace CorePayAPI.Entities
-{
-    public class Entity : IEntity
-    {
-        public int Id { get; set; }
-    }
-}
