--- CorePayAPI/Entities/Interface/IEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace CorePayAPI.Entities.Interface
-{
-    public interface IEntity
-    {
-        public int Id { get; set; }
-    }
-}
