--- CorePayAPI/Repository/Interface/IUserRepository.cs
@@ -1,10 +1,0 @@
-﻿using CorePayAPI.Entities.CorePayDB;
-
-namespace CorePayAPI.Repository.Interface
-{
-    public interface IUserRepository
-    {
-        User ConsultUser(int UserId);
-
-    }
-}
