--- CorePayAPI/Services/Interface/IUserService.cs
@@ -1,9 +1,0 @@
-﻿using CorePayAPI.Entities.CorePayDB;
-
-namespace CorePayAPI.Services.Interface
-{
-    public interface IUserService
-    {
-        User ConsultUser(int userId);
-    }
-}
