+++ CorePayAPI/appsettings.json
@@ -6,7 +6,7 @@
     }
   },
   "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "databaseUrl": "Server=(localdb)\\MSSQLLocalDB;Database=CorePayDB;Trusted_Connection=True;"
+  "DbConfig": {
+    "ConnectionString": "Server=localhost,1433;Database=CorePayDb;User Id=sa;Password=StrongPass!123;TrustServerCertificate=true;"
   }
 }
